conversao sem float

+++ conversao.c
@@ -4,7 +4,7 @@ int main() {
     int a = 10;
     int b = 3;
 
-    float quociente = (float) a / b; // conversão explícita
+    float quociente = a / b; // conversão explícita
 
     printf("Quociente: %.2f\n", quociente);
